pyproject.toml
[tool.poetry]
name = "onlinestore"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
django-filter = "^25.1"
pillow = "^11.3.0"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
cloudinary = "^1.44.1"
django-cloudinary-storage = "^0.3.0"
gunicorn = "^23.0.0"
psycopg = "^3.2.9"
pytest = "^7.4.0"
pytest-django = "^4.5.2"
django-import-export-extensions = "^1.8.0"
django-import-export = "^4.3.9"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
ruff = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "configs.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
